+++ src/main/java/com/atlassian/db/replica/internal/ReplicaStatement.java
@@ -37,6 +37,7 @@ public class ReplicaStatement implements Statement {
             return createStatement(connectionProvider.getWriteConnection(getFirstCause()));
         }
     };
+    private final boolean compatibleWithPreviousVersion;
 
     public ReplicaStatement(
         ReplicaConsistency consistency,
@@ -45,7 +46,8 @@ public class ReplicaStatement implements Statement {
         Integer resultSetType,
         Integer resultSetConcurrency,
         Integer resultSetHoldability,
-        Set<String> readOnlyFunctions
+        Set<String> readOnlyFunctions,
+        boolean compatibleWithPreviousVersion
     ) {
         this.consistency = consistency;
         this.connectionProvider = connectionProvider;
@@ -54,6 +56,7 @@ public class ReplicaStatement implements Statement {
         this.resultSetConcurrency = resultSetConcurrency;
         this.resultSetHoldability = resultSetHoldability;
         this.sqlFunction = new SqlFunction(readOnlyFunctions);
+        this.compatibleWithPreviousVersion = compatibleWithPreviousVersion;
     }
 
     @Override
@@ -175,7 +178,7 @@ public class ReplicaStatement implements Statement {
         checkClosed();
         final RouteDecisionBuilder decisionBuilder;
         final Statement statement;
-        SqlQuery sqlQuery = new SqlQuery(sql);
+        SqlQuery sqlQuery = new SqlQuery(sql, compatibleWithPreviousVersion);
         if (sqlQuery.isSqlSet()) {
             decisionBuilder = new RouteDecisionBuilder(READ_OPERATION).sql(sql);
             statement = getReadStatement(decisionBuilder);
@@ -540,9 +543,12 @@ public class ReplicaStatement implements Statement {
         ReplicaConnectionProvider connectionProvider,
         ReplicaConsistency consistency,
         DatabaseCall databaseCall,
-        Set<String> readOnlyFunctions
+        Set<String> readOnlyFunctions,
+        boolean compatibleWithPreviousVersion
     ) {
-        return new Builder(connectionProvider, consistency, databaseCall, readOnlyFunctions);
+        return new Builder(connectionProvider, consistency, databaseCall, readOnlyFunctions,
+            compatibleWithPreviousVersion
+        );
     }
 
     void recordWriteAfterQueryExecution() throws SQLException {
@@ -558,7 +564,7 @@ public class ReplicaStatement implements Statement {
             connectionProvider.getStateDecision().ifPresent(decisionBuilder::cause);
             return prepareWriteStatement(decisionBuilder);
         }
-        SqlQuery sqlQuery = new SqlQuery(decisionBuilder.getSql());
+        SqlQuery sqlQuery = new SqlQuery(decisionBuilder.getSql(), compatibleWithPreviousVersion);
         if (sqlQuery.isWriteOperation(sqlFunction)) {
             decisionBuilder.reason(WRITE_OPERATION);
             return prepareWriteStatement(decisionBuilder);
@@ -609,6 +615,7 @@ public class ReplicaStatement implements Statement {
         private final ReplicaConsistency consistency;
         private final DatabaseCall databaseCall;
         private final Set<String> readOnlyFunctions;
+        private final boolean compatibleWithPreviousVersion;
         private Integer resultSetType;
         private Integer resultSetConcurrency;
         private Integer resultSetHoldability;
@@ -617,12 +624,14 @@ public class ReplicaStatement implements Statement {
             ReplicaConnectionProvider connectionProvider,
             ReplicaConsistency consistency,
             DatabaseCall databaseCall,
-            Set<String> readOnlyFunctions
+            Set<String> readOnlyFunctions,
+            boolean compatibleWithPreviousVersion
         ) {
             this.connectionProvider = connectionProvider;
             this.consistency = consistency;
             this.databaseCall = databaseCall;
             this.readOnlyFunctions = readOnlyFunctions;
+            this.compatibleWithPreviousVersion = compatibleWithPreviousVersion;
         }
 
         public Builder resultSetType(int resultSetType) {
@@ -648,7 +657,8 @@ public class ReplicaStatement implements Statement {
                 resultSetType,
                 resultSetConcurrency,
                 resultSetHoldability,
-                readOnlyFunctions
+                readOnlyFunctions,
+                compatibleWithPreviousVersion
             );
         }
     }